backend/vehicles/urls.py
from django.urls import include, path
from rest_framework.routers import DefaultRouter

from vehicles.views import VehicleViewSet

router = DefaultRouter()
router.register(r"vehicles", VehicleViewSet, basename="vehicle")

urlpatterns = [path("", include(router.urls))]
