backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
django-cors-headers==4.9.0
psycopg2-binary==2.9.12
celery==5.6.3
redis==7.4.0
python-dotenv==1.2.2
